Version 1
ResNet 72 with Xavier initialization for Linear and Conv2D and Normal initializatopn for BN layers. the preformence are still very poor, the network size is 22M parameters. for conclusion, the images don't contain enpought data to conclude anything from them.  Maybe in the future we will try different archtictures, but maenwhile, I'm going to next project.